Pull out class to separate file, fix lint warnings
